Rename `type rr` to `type response`
The current definition is awkward because it has 3 optional fields
which are `None` when we are unable to parse the resource record.
This patch renames the `type rr` to type `response` i.e. the result
is a "response" even if we aren't sure if it's a valid "resource
record". A future patch will coalesce the `None` cases together.

Signed-off-by: David Scott <dave@recoil.org>