Recursively lookup CNAMES to A/AAAA/...
Previously the raw binding would only return CNAMEs and the user would
have to recursively query it themselves. This patch performs the
recursive lookup inside the library to hide this quirk from users.

Note the NoSuchRecord handling: if we query a name which is a CNAME
which maps to a missing record, we should return the CNAME and treat
the lookup as a success. We only return NoSuchRecord if we found no
records at all.

Signed-off-by: David Scott <dave.scott@docker.com>